Address combobox review feedback
