+++ packages/components/src/lib/native-popover-sizing.ts
@@ -1,0 +1,22 @@
+const NATIVE_POPOVER_MAX_HEIGHT = 300;
+const NATIVE_POPOVER_VERTICAL_PADDING = 16;
+const NATIVE_SEARCH_INPUT_HEIGHT_WITH_MARGIN = 48;
+const NATIVE_LIST_MIN_ITEM_HEIGHT = 36;
+const NATIVE_LIST_MAX_HEIGHT =
+	NATIVE_POPOVER_MAX_HEIGHT -
+	NATIVE_POPOVER_VERTICAL_PADDING -
+	NATIVE_SEARCH_INPUT_HEIGHT_WITH_MARGIN;
+
+function getNativeListHeight(itemCount: number, estimatedItemSize: number) {
+	const itemHeight = Math.max(estimatedItemSize, NATIVE_LIST_MIN_ITEM_HEIGHT);
+	return Math.min(itemCount * itemHeight, NATIVE_LIST_MAX_HEIGHT);
+}
+
+export {
+	NATIVE_LIST_MAX_HEIGHT,
+	NATIVE_LIST_MIN_ITEM_HEIGHT,
+	NATIVE_POPOVER_MAX_HEIGHT,
+	NATIVE_POPOVER_VERTICAL_PADDING,
+	NATIVE_SEARCH_INPUT_HEIGHT_WITH_MARGIN,
+	getNativeListHeight,
+};
