+++ packages/components/src/tree-combobox/tree-combobox.tsx
@@ -13,6 +13,11 @@ import { Icon } from '../icon';
 import { Input } from '../input';
 import { useArrowKeyNavigation } from '../lib/use-arrow-key-navigation';
 import { applyCascadeToggle, useHierarchy } from '../lib/use-hierarchy';
+import {
+	getNativeListHeight,
+	NATIVE_LIST_MAX_HEIGHT,
+	NATIVE_POPOVER_MAX_HEIGHT,
+} from '../lib/native-popover-sizing';
 import { cn } from '../lib/utils';
 import { Text, TextClassContext } from '../text';
 import { INDENT_PX } from '../tree-select/tree-item';
@@ -22,20 +27,6 @@ import type { FlatTreeItem } from '../lib/use-hierarchy';
 import type { TreeComboboxContentProps, TreeComboboxProps } from './types';
 
 type ComboboxOption<T = undefined> = { value: string; label: string; item?: T };
-
-const NATIVE_POPOVER_MAX_HEIGHT = 300;
-const NATIVE_POPOVER_VERTICAL_PADDING = 16;
-const NATIVE_SEARCH_INPUT_HEIGHT_WITH_MARGIN = 48;
-const NATIVE_LIST_MIN_ITEM_HEIGHT = 36;
-const NATIVE_LIST_MAX_HEIGHT =
-	NATIVE_POPOVER_MAX_HEIGHT -
-	NATIVE_POPOVER_VERTICAL_PADDING -
-	NATIVE_SEARCH_INPUT_HEIGHT_WITH_MARGIN;
-
-function getNativeListHeight(itemCount: number, estimatedItemSize: number) {
-	const itemHeight = Math.max(estimatedItemSize, NATIVE_LIST_MIN_ITEM_HEIGHT);
-	return Math.min(itemCount * itemHeight, NATIVE_LIST_MAX_HEIGHT);
-}
 
 // --- Context ---
